test/perceptron-test.html

<!DOCTYPE html>
<html>
<head>
	<title>Perceptron Tests</title>
	<script src="../perceptron.js"></script>
	<script src="testharness.js"></script>
	<script src="testharnessreport.js"></script>
</head>
<body>
	<script>
		test(function() {
			var initialWeights = [1,2,3];
			var perceptron = new Perceptron(initialWeights);
			assert_equals(perceptron.weights.length, initialWeights.length);

			for (var i = 0; i < initialWeights.length; i++) {
				assert_equals(perceptron.weights[i], initialWeights[i]);
			}
		}, "initial weights accepted");

		test(function() {
			var initialWeights = [1,2,3];
			var classifierFunction = function(weightedSum) {
				return weightedSum >= 0 ? +1 : -1;
			}

			var perceptron = new Perceptron(initialWeights, classifierFunction);
			assert_equals(perceptron.classifierFunction, classifierFunction);
		}, "classifier function accepted");

		test(function() {
			var moreThanFivePerceptron = new Perceptron(/*initial weights*/[0,0]);

			var maxSteps = 1600; // range of the domain squared over the allowed error squared iterations is guaranteed to converge
			for (var step = 0; step < maxSteps; step++) {
				// compute random stimulus
				var sign = Math.random() > 0.5 ? +1 : -1;
				var magnitude = Math.random() * 10;
				stimulus = sign * magnitude;
				
				// determine the desired output
				desiredResponse = stimulus > 5 ? +1 : -1;

				// learn
				moreThanFivePerceptron.learn([stimulus], desiredResponse);

				// for this perceptron we need a positive slope and an x-intercept close to five for it to have converged on a solution
				var m = moreThanFivePerceptron.weights[0]; // slope
				var b = moreThanFivePerceptron.weights[1]; // y-intercept
				var x = -b / m - 5; // x-intercept
				var allowedError = 0.5;

				if (m > 0 && x < allowedError && x > -allowedError) {
					break;
				}
			}

			console.log(
				" weight0:" + moreThanFivePerceptron.weights[0] +
				" bias:" + moreThanFivePerceptron.weights[1]
			);

			assert_true(step < maxSteps, "perceptron did not converge on its weights and learned bias in " + maxSteps + " steps");

		}, "Perceptron converged to classify the set of numbers greater than 5");

	</script>
	<div id="log"></div>
</body>
</html>
